move nutrition total calculation to its own script
